requirements.txt
minio==4.0.18
pika==1.0.1
mongo